Update DOTHUB_LUMO2nirs.m
Added handling of zero-valued intensity values (i.e. they are converted to the noise floor estimate, being the mean of channels with an SD distance greater than or equal to 70 mm. A warning message is displayed if the conversion occurs, or if maximum SD distance is less than 70 mm and the data still contain zero values.