Update DOTHUB_plotIntVTime.m
Might it be useful to include the wavelength of the channel in the label? I have updated the function DOTHUB_LUMOintvtimeplot_tiplabels to do this. It works on the assumption that the first half of channels are SD.lambda(1) and the second half of channels are SD.lambda(2) for LUMO data