Update DOTHUB_plotPSD.m
line 50 should read (3/4) not (2/3) - think this is a bug? i.e. there are 3 sources for every 4 detectors 
I've updated the function to make a call to DOTHUB_getSDdists too, as this is required in line 50
and I've updated DOTHUB_LUMOpsdplot_tiplabels to show wavelength information too on the active label